fix: tests and circular deps
