+++ packages/apps/reputation-oracle/server/src/common/enums/index.ts
@@ -1,4 +1,5 @@
 export * from './collection';
 export * from './http';
 export * from './manifest';
+export * from './user';
 export * from './web3';
