+++ packages/apps/reputation-oracle/server/src/database/database.module.ts
@@ -4,24 +4,23 @@ import { LoggerOptions } from 'typeorm';
 import { SnakeNamingStrategy } from 'typeorm-naming-strategies';
 
 import { DatabaseConfigService } from '@/config';
-import { AbuseEntity } from '@/modules/abuse';
-import { TokenEntity } from '@/modules/auth';
-import { CronJobEntity } from '@/modules/cron-job';
-import {
-  EscrowCompletionEntity,
-  EscrowPayoutsBatchEntity,
-} from '@/modules/escrow-completion';
-import { KycEntity } from '@/modules/kyc';
-import {
-  QualificationEntity,
-  UserQualificationEntity,
-} from '@/modules/qualification';
-import { ReputationEntity } from '@/modules/reputation';
-import { UserEntity, SiteKeyEntity } from '@/modules/user';
-import {
-  IncomingWebhookEntity,
-  OutgoingWebhookEntity,
-} from '@/modules/webhook';
+/**
+ * Entities should be imported directly to avoid circular dependency
+ * between entity module and DB module
+ */
+import { AbuseEntity } from '@/modules/abuse/abuse.entity';
+import { TokenEntity } from '@/modules/auth/token.entity';
+import { CronJobEntity } from '@/modules/cron-job/cron-job.entity';
+import { EscrowCompletionEntity } from '@/modules/escrow-completion/escrow-completion.entity';
+import { EscrowPayoutsBatchEntity } from '@/modules/escrow-completion/escrow-payouts-batch.entity';
+import { KycEntity } from '@/modules/kyc/kyc.entity';
+import { QualificationEntity } from '@/modules/qualification/qualification.entity';
+import { UserQualificationEntity } from '@/modules/qualification/user-qualification.entity';
+import { ReputationEntity } from '@/modules/reputation/reputation.entity';
+import { SiteKeyEntity } from '@/modules/user/site-key.entity';
+import { UserEntity } from '@/modules/user/user.entity';
+import { IncomingWebhookEntity } from '@/modules/webhook/webhook-incoming.entity';
+import { OutgoingWebhookEntity } from '@/modules/webhook/webhook-outgoing.entity';
 
 import { TypeOrmLoggerModule, TypeOrmLoggerService } from './typeorm';
