+++ packages/apps/reputation-oracle/server/src/modules/auth/dto/sign-up.dto.ts
@@ -1,12 +1,12 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsEthereumAddress, IsString } from 'class-validator';
 
+import { UserRole } from '@/common/enums';
 import {
   IsLowercasedEmail,
   IsLowercasedEnum,
   IsValidWeb3Signature,
 } from '@/common/validators';
-import { UserRole } from '@/modules/user';
 
 import { ValidPassword } from './password.dto';
