+++ packages/apps/reputation-oracle/server/src/modules/kyc/index.ts
@@ -1,3 +1,2 @@
 export { KycModule } from './kyc.module';
 export { KycEntity } from './kyc.entity';
-export { KycStatus } from './constants';
