+++ packages/apps/reputation-oracle/server/src/modules/qualification/qualification.service.spec.ts
@@ -3,8 +3,9 @@ import { createMock } from '@golevelup/ts-jest';
 import { ConfigService } from '@nestjs/config';
 import { Test } from '@nestjs/testing';
 
+import { UserStatus } from '@/common/enums';
 import { ServerConfigService } from '@/config';
-import { UserStatus, UserRepository } from '@/modules/user';
+import { UserRepository } from '@/modules/user';
 import { generateWorkerUser } from '@/modules/user/fixtures';
 import { generateFutureDate } from '~/test/fixtures/date';
 import { generateEthWallet } from '~/test/fixtures/web3';
