+++ packages/apps/reputation-oracle/server/src/modules/user/index.ts
@@ -1,7 +1,7 @@
 export { SiteKeyRepository } from './site-key.repository';
 export { SiteKeyType, SiteKeyEntity } from './site-key.entity';
 export type * from './types';
-export { UserEntity, UserStatus, Role as UserRole } from './user.entity';
+export { UserEntity } from './user.entity';
 export { UserNotFoundError } from './user.error';
 export { UserModule } from './user.module';
 export { UserRepository } from './user.repository';
