+++ packages/apps/reputation-oracle/server/src/modules/user/user.service.ts
@@ -2,16 +2,15 @@ import { KVStoreClient, KVStoreUtils } from '@human-protocol/sdk';
 import { Injectable } from '@nestjs/common';
 
 import { SignatureType } from '@/common/enums';
+import { KycStatus, UserRole } from '@/common/enums';
 import { Web3ConfigService } from '@/config';
 import { HCaptchaService } from '@/integrations/hcaptcha/hcaptcha.service';
-import { KycStatus } from '@/modules/kyc';
 import { Web3Service } from '@/modules/web3';
 import * as web3Utils from '@/utils/web3';
 
 import { SiteKeyEntity, SiteKeyType } from './site-key.entity';
 import { SiteKeyRepository } from './site-key.repository';
 import { OperatorUserEntity, Web2UserEntity } from './types';
-import { Role as UserRole } from './user.entity';
 import {
   DuplicatedWalletAddressError,
   InvalidWeb3SignatureError,
